Fail and correctly parse on any error in the requester.
This matches the behavior of the Python and Go SDKs. It also fixes some error parsing issues when the top level data is non-null, but the only useful fields are. For example, when querying Account.getTransactions() with a page size >200, it should fire an error with type MaxDatabaseQueryException with a reasonable message. However, previously, it was firing an opaque json deserialization error.